% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashboardBuildR-package.R
\docType{package}
\name{dashboardBuildR-package}
\alias{dashboardBuildR}
\alias{dashboardBuildR-package}
\title{dashboardBuildR}
\description{
Tools and a Shiny application for building NISRA data dashboards.
}
\details{
The main user-facing entry point is
\code{\link[=run_dashboard_buildr]{run_dashboard_buildr()}}.
}
\author{
\strong{Maintainer}: NISRA Tech Lab \email{nisra.statistics@finance-ni.gov.uk}

Authors:
\itemize{
  \item NISRA Tech Lab \email{nisra.statistics@finance-ni.gov.uk}
}

}
\keyword{internal}
