% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.R
\name{run_dashboard_buildr}
\alias{run_dashboard_buildr}
\title{Run Dashboard BuildR}
\usage{
run_dashboard_buildr()
}
\value{
No return value. Launches the Shiny application.
}
\description{
Launch the Dashboard BuildR Shiny application in the default web browser.
}
